Turned fastwave example into a test
